refactor: encapsulate vendor fetch to CVEDB
