+++ cve_bin_tool/cvedb.py
@@ -718,6 +718,47 @@ class CVEDB:
             self.LOGGER.warning(f"Deleting old cachedir {OLD_CACHE_DIR}")
             shutil.rmtree(OLD_CACHE_DIR)
 
+    def get_vendor_product_pairs(self, package_names):
+        """
+        Fetches vendor from the database for packages that doesn't have vendor info for Package List Parser Utility and Universal Python package checker.
+        """
+        self.db_open()
+        cursor = self.connection.cursor()
+        vendor_package_pairs = []
+        query = """
+        SELECT DISTINCT vendor FROM cve_range
+        WHERE product=?
+        """
+
+        # For python package checkers we don't need the progress bar running
+        if type(package_names) != list:
+            cursor.execute(query, [package_names])
+            vendors = list(map(lambda x: x[0], cursor.fetchall()))
+            for vendor in vendors:
+                if vendor != "":
+                    vendor_package_pairs.append(
+                        {
+                            "vendor": vendor,
+                            "product": package_names,
+                        }
+                    )
+        else:
+            for package_name in track(
+                package_names, description="Processing the given list...."
+            ):
+                cursor.execute(query, [package_name["name"].lower()])
+                vendors = list(map(lambda x: x[0], cursor.fetchall()))
+                for vendor in vendors:
+                    if vendor != "":
+                        vendor_package_pairs.append(
+                            {
+                                "vendor": vendor,
+                                "product": package_name["name"],
+                            }
+                        )
+        self.db_close()
+        return vendor_package_pairs
+
     def db_open(self):
         """Opens connection to sqlite database."""
         if not self.connection:
